coffee counter
In this example, the data variable stores the number of visitors to the coffee shop at each hour. The code then calculates the average number of visitors during each hour, and identifies the hour(s) with the highest average count. Finally, it outputs the busiest hour(s) to the console.